+++ pkgs/test_core/lib/src/runner/configuration/args.dart
@@ -102,9 +102,15 @@ final ArgParser _parser = (() {
           'especially for asynchronous code. It may be useful to disable\n'
           'to provide improved test performance but at the cost of\n'
           'debuggability.',
-      defaultsTo: true);
+      defaultsTo: false);
   parser.addFlag('no-retry',
       help: "Don't rerun tests that have retry set.",
+      defaultsTo: false,
+      negatable: false);
+  parser.addFlag('use-data-isolate-strategy',
+      help: 'Use `data:` uri isolates when spawning VM tests instead of the\n'
+          'default strategy. This may be faster when you only ever run a\n'
+          'single test suite at a time.',
       defaultsTo: false,
       negatable: false);
   parser.addOption('test-randomize-ordering-seed',
@@ -230,15 +236,6 @@ class _Parser {
     var platform = _ifParsed<List<String>>('platform')
         ?.map((runtime) => RuntimeSelection(runtime))
         .toList();
-    if (platform
-            ?.any((runtime) => runtime.name == Runtime.phantomJS.identifier) ??
-        false) {
-      var yellow = color ? '\u001b[33m' : '';
-      var noColor = color ? '\u001b[0m' : '';
-      print('${yellow}Warning:$noColor '
-          'PhatomJS is deprecated and will be removed in version ^2.0.0');
-    }
-
     return Configuration(
         help: _ifParsed('help'),
         version: _ifParsed('version'),
@@ -268,6 +265,7 @@ class _Parser {
         includeTags: includeTags,
         excludeTags: excludeTags,
         noRetry: _ifParsed('no-retry'),
+        useDataIsolateStrategy: _ifParsed('use-data-isolate-strategy'),
         testRandomizeOrderingSeed: testRandomizeOrderingSeed);
   }