+++ pkgs/test_core/lib/src/runner/hybrid_listener.dart
@@ -54,8 +54,17 @@ void listen(Function Function() getMain, List data) {
         return;
       } else if (main is! Function(StreamChannel) &&
           main is! Function(StreamChannel, Never)) {
-        _sendError(channel,
-            'Top-level hybridMain() function must take one or two arguments.');
+        if (main is Function(StreamChannel<Never>) ||
+            main is Function(StreamChannel<Never>, Never)) {
+          _sendError(
+              channel,
+              'The first parameter to the top-level hybridMain() must be a '
+              'StreamChannel<dynamic> or StreamChannel<Object?>. More specific '
+              'types such as StreamChannel<Object> are not supported.');
+        } else {
+          _sendError(channel,
+              'Top-level hybridMain() function must take one or two arguments.');
+        }
         return;
       }