+++ pkgs/test_api/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.4.0-dev
+
+* Add examples to `throwsA` and make top-level `throws...` matchers refer to it.
+* Disable stack trace chaining by default.
+* **Breaking** remove `Runtime.phantomJS`
+
 ## 0.3.0
 
 * **Breaking** `TestException.message` is now nullable.