+++ pkgs/test_core/CHANGELOG.md
@@ -1,3 +1,17 @@
+## 0.3.20-dev
+
+* Disable stack trace chaining by default.
+* Change the default way VM tests are launched and ran to greatly speed up
+  loading performance.
+  * You can force the old strategy with `--use-data-isolate-strategy` flag if
+    you run into issues, but please also file a bug.
+* Improve the error message for `hybridMain` functions with an incompatible
+  StreamChannel parameter type.
+
+## 0.3.19
+
+* ~~Disable stack trace chaining by default.~~
+
 ## 0.3.18
 
 * Update `spawnHybridCode` to default to the current packages language version.